requirements.txt
cdmkdocs-material==9.4.7

Django~=4.2.7
djangorestframework~=3.14.0